Logging is off by default
It can be turned on and off freely. When on, the log file is written
and can be viewed in the app. When disaling, the log file is removed.

Upgrading users need to enable and disable logging to remove previously
written logs.

This change also tweaks a bunch of terms for consistency.